Attempted to fix configkv plugin (did not work).
I tried resolve this bug, but does not work, have to revisit. Saving commit for now.
The bug is about `default_` properties profiles that contain a property default such as, say, ${project.basedir}, which is convenient to use for absolute paths (user does not have to necessarily put that in settigns.xml).
If the user does so, that property ends up being listed in the defaults sectino of the output file of this plugin (configkv.properties), but it cannot possibly be a default, because it has an absolute path.

Unfortunately, when i try distunguish between a property that has a token or not, it is too late, because Maven has already resolved it. Need to figure out how to parse the pom and see if the value textually contains a `$` symbol or not.